+++ src/sse/Backend/GoogleBackend.java
@@ -55,7 +55,7 @@ public class GoogleBackend implements Backend {
 	 * @throws IOException
 	 */
 	@Override
-	public DataOutputStream openBlock(int block) {
+	public DataOutputStream openBlock(long block) {
 		try {
 			w.close();
 		} catch (IOException e1) {
