+++ src/sse/Backend/SmartdriveBackend.java
@@ -56,7 +56,7 @@ public class SmartdriveBackend implements Backend {
 	}
 
 	@Override
-	public DataOutputStream openBlock(int block) {
+	public DataOutputStream openBlock(long block) {
 
 		if (w != null) {
 			try {
